+++ ft-derive/src/lib.rs
@@ -32,50 +32,6 @@ pub fn form(
     handle(item, "form", "handler")
 }
 
-#[proc_macro_attribute]
-pub fn migration(
-    _attr: proc_macro::TokenStream,
-    item: proc_macro::TokenStream,
-) -> proc_macro::TokenStream {
-    let syn::ItemFn {
-        attrs,
-        vis,
-        sig,
-        block,
-    } = syn::parse_macro_input!(item as syn::ItemFn);
-
-    let fn_name = &sig.ident;
-
-    let expanded = quote::quote! {
-        #[no_mangle]
-        pub extern "C" fn migration__entrypoint() -> i32 { // return 0 for success
-            let mut conn = match ft_sdk::default_connection() {
-                Ok(c) => c,
-                Err(e) => {
-                    ft_sdk::println!("error when getting connection to apply migration: {e}");
-                    return 1;
-                }
-            };
-
-            match #fn_name(conn) {
-                Ok(()) => 0,
-                Err(e) => {
-                    ft_sdk::println!("error when applying migration: {e}");
-                    1
-                }
-            }
-        }
-
-        #(#attrs)*
-        #vis #sig {
-            #block
-        }
-    };
-
-    // println!("{expanded}");
-    proc_macro::TokenStream::from(expanded)
-}
-
 fn handle(item: proc_macro::TokenStream, kind: &str, handler: &str) -> proc_macro::TokenStream {
     let syn::ItemFn {
         attrs,