+++ ft-sdk/src/lib.rs
@@ -17,10 +17,6 @@ mod email;
 mod error;
 pub mod form;
 pub mod from_request;
-#[cfg(all(
-    feature = "migration",
-    any(feature = "postgres-default", feature = "sqlite-default")
-))]
 pub mod processor;
 mod rng;
 pub mod utils;
@@ -33,7 +29,7 @@ pub use error::{not_found_, server_error_, single_error, unauthorised_, SpecialE
 #[cfg(feature = "field-extractors")]
 pub use from_request::{Cookie, Hidden, Optional, Query, Required};
 pub use from_request::{Form, FromRequest, Host, Mountpoint, Path, WrappedFromRequest};
-pub use ft_derive::{data, form, migration, processor, wrapped_processor};
+pub use ft_derive::{data, form, processor, wrapped_processor};
 #[cfg(feature = "postgres")]
 pub use ft_sys::PgConnection;
 #[cfg(feature = "sqlite")]